+++ Specs/Core/PolylinePipelineSpec.js
@@ -8,7 +8,7 @@ defineSuite([
          Cartographic,
          Ellipsoid) {
     "use strict";
-    /*global it,expect*/
+    /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
     it('wrapLongitude', function() {
         var ellipsoid = Ellipsoid.WGS84;